Restructure Project
This commit restructures the base project and adds a few features
as well. Dependencies have also been updated.

+ src/
This is where all of the application's code will live. The "source"
that will then be transpiled by Babel.

+ dist/
This is the generated distribution directory where finished and
compiled code will live (not pushed to git).

+ .*ignore
All ignore files have been updated so as not to accidentally leak
any files that don't belong on a particular system.

+ Dockerfile.*
New Dockerfiles have been created for both development and
production. I'll be adding a compose file very soon.

+ package.json
Updated dependencies and npm scripts!